pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
lxml = "^5.1.0"
beautifulsoup4 = "^4.12.3"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.27"
openai = "^1.12.0"
aiogram = "^3.4.1"
aiohttp-requests = "^0.2.3"
apscheduler = "^3.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
